requirements.txt
ipython
jupyterlab
loguru
matplotlib
mkdocs
notebook
numpy
pandas
seaborn
pip
python-dotenv
ruff
scikit-learn
tqdm
typer
spacy
nltk
psycopg2-binary
pgvector
ollama
sentence-transformers
